Add files via upload
This file talks about very basic data exploration, through the loading of dataset into pandas, making a usable object in python and then using the .describe() method to get a summary of the key averages and statistics. We then talk about the dataset's timeline and quality, potential drawbacks and advantages.